I3412 added an explicit conversion

+++ src/fheroes2/heroes/heroes.cpp
@@ -399,7 +399,7 @@ Army & Heroes::GetArmy( void )
 int Heroes::GetMobilityIndexSprite( void ) const
 {
     // valid range (0 - 25)
-    int index = !CanMove() ? 0 : std::round( move_point / 100.0f );
+    int index = !CanMove() ? 0 : static_cast<int>(std::round( move_point / 100.0f ));
     return 25 >= index ? index : 25;
 }
